refactor: separate SQLite CRUD from HistoryViewModel
Extracts SQLite database, CRUD, and migration logic into HistoryDatabaseManager. Updates CloudKitSyncManager and HistoryViewModel to use the new database manager while keeping UI/view logic in HistoryViewModel.
