+++ Source/Managers/ViewModels/HistoryViewModel.swift
@@ -46,27 +46,19 @@ class HistoryViewModel: ViewModelBase, ObservableObject {
 
     private let maxHistoryCount = 50
 
-    /// Legacy UserDefaults keys — used only for migration
-    private let legacyStorageKey = "CloudReadingEntries"
+    /// Legacy UserDefaults keys
     private let legacyHistoryKey = "iOSReadingEntries"
-    private let legacyPendingUploadsKey = "HistoryPendingUploads"
-    private let legacyPendingDeletesKey = "HistoryPendingDeletes"
-    private let migrationFlag = "HistoryVM_SQLiteMigrated"
 
     /// Debounce for batched CloudKit deletions
     private var pendingCloudKitDeletes: Set<String> = []
     private var deleteDebounceTask: Task<Void, Never>?
-
-    // MARK: - Database
-
-    private var _db: SQLiteDatabase?
 
     var historyBookIds: [Int] {
         get { historyOrder }
         set {
             historyOrder = Array(newValue.prefix(maxHistoryCount))
             pruneOrphanedEntries()
-            saveHistoryOrder()
+            HistoryDatabaseManager.shared.saveHistoryOrder(historyOrder)
             loadBooksData()
         }
     }
@@ -88,8 +80,8 @@ class HistoryViewModel: ViewModelBase, ObservableObject {
 
     override init() {
         super.init()
-        setupDatabase()
-        migrateFromUserDefaultsIfNeeded()
+        HistoryDatabaseManager.shared.setupDatabase()
+        HistoryDatabaseManager.shared.migrateFromUserDefaultsIfNeeded()
         migrateLegacyKVSDataIfNeeded()
         loadFromDatabase()
         backfillCloudKitFieldsIfNeeded()
@@ -116,155 +108,12 @@ class HistoryViewModel: ViewModelBase, ObservableObject {
         }
     }
 
-    // MARK: - Database Setup
-
-    private func setupDatabase() {
-        guard let folderURL = AppConfig.folder(for: AppConfig.annotationsAndResultsFolder) else {
-            #if DEBUG
-            print("HistoryViewModel: No folder URL available for History database")
-            #endif
-            return
-        }
-
-        let fm = FileManager.default
-        if !fm.fileExists(atPath: folderURL.path) {
-            try? fm.createDirectory(at: folderURL, withIntermediateDirectories: true)
-        }
-
-        let url = folderURL.appendingPathComponent("History.sqlite")
-
-        do {
-            _db = try SQLiteDatabase(path: url.path)
-            enableWALMode()
-            try createTables()
-        } catch {
-            #if DEBUG
-            print("HistoryViewModel: Failed to setup database: \(error)")
-            #endif
-        }
-    }
-
-    private func enableWALMode() {
-        guard let _db else { return }
-        do {
-            let mode = try _db.fetch(query: "PRAGMA journal_mode = WAL;") { row in
-                row.string(at: 0) ?? ""
-            }.first
-
-            #if DEBUG
-            if mode?.lowercased() != "wal" {
-                print("HistoryViewModel: failed to enable WAL mode, current: \(mode ?? "unknown")")
-            }
-            #endif
-        } catch {
-            #if DEBUG
-            print("HistoryViewModel: error enabling WAL mode: \(error)")
-            #endif
-        }
-    }
-
-    private func createTables() throws {
-        try exec("""
-        CREATE TABLE IF NOT EXISTS reading_entries (
-            book_id INTEGER PRIMARY KEY,
-            last_content_id INTEGER,
-            last_opened_at REAL,
-            favorited_at REAL,
-            position_updated_at REAL,
-            updated_at REAL NOT NULL,
-            is_favorite INTEGER NOT NULL DEFAULT 0,
-            ck_record_id TEXT
-        );
-        """)
-
-        try exec("""
-        CREATE TABLE IF NOT EXISTS history_order (
-            position INTEGER PRIMARY KEY,
-            book_id INTEGER NOT NULL
-        );
-        """)
-
-        try exec("""
-        CREATE TABLE IF NOT EXISTS sync_pending (
-            ck_record_id TEXT PRIMARY KEY,
-            operation TEXT NOT NULL CHECK(operation IN ('upload', 'delete')),
-            queued_at INTEGER NOT NULL
-        );
-        """)
-
-        try exec("CREATE INDEX IF NOT EXISTS idx_re_favorite ON reading_entries (is_favorite, favorited_at);")
-        try exec("CREATE INDEX IF NOT EXISTS idx_sync_pending_op ON sync_pending (operation, queued_at);")
-    }
-
-    // MARK: - SQLite Helpers
-
-    private func exec(_ sql: String, parameters: [Any] = []) throws {
-        guard let _db else { return }
-        try _db.execute(query: sql, parameters: parameters)
-    }
-
-    private func transaction(_ block: () throws -> Void) throws {
-        guard let _db else { return }
-        try _db.transaction(block)
-    }
-
-    // MARK: - Core CRUD
-
-    private func upsertEntry(_ entry: ReadingEntry) {
-        let sql = """
-        INSERT OR REPLACE INTO reading_entries
-        (book_id, last_content_id, last_opened_at, favorited_at, position_updated_at, updated_at, is_favorite, ck_record_id)
-        VALUES (?, ?, ?, ?, ?, ?, ?, ?);
-        """
-        let params: [Any] = [
-            entry.bookId,
-            entry.lastContentId as Any? ?? NSNull(),
-            entry.lastOpenedAt?.timeIntervalSince1970 as Any? ?? NSNull(),
-            entry.favoritedAt?.timeIntervalSince1970 as Any? ?? NSNull(),
-            entry.positionUpdatedAt?.timeIntervalSince1970 as Any? ?? NSNull(),
-            entry.updatedAt.timeIntervalSince1970,
-            entry.isFavorite ? 1 : 0,
-            entry.ckRecordId as Any? ?? NSNull(),
-        ]
-        try? _db?.execute(query: sql, parameters: params)
-    }
-
-    private func deleteEntry(bookId: Int) {
-        try? exec("DELETE FROM reading_entries WHERE book_id = ?;", parameters: [bookId])
-    }
-
-    private func saveHistoryOrder() {
-        try? transaction {
-            try exec("DELETE FROM history_order;")
-            for (position, bookId) in historyOrder.enumerated() {
-                try exec("INSERT INTO history_order (position, book_id) VALUES (?, ?);", parameters: [position, bookId])
-            }
-        }
-    }
-
     // MARK: - Load from Database
 
     private func loadFromDatabase() {
-        guard let _db else { return }
-
-        let entries = (try? _db.fetch(query: "SELECT book_id, last_content_id, last_opened_at, favorited_at, position_updated_at, updated_at, is_favorite, ck_record_id FROM reading_entries;") { row -> ReadingEntry in
-            ReadingEntry(
-                bookId: row.int(at: 0),
-                lastContentId: row.isNull(at: 1) ? nil : row.int(at: 1),
-                lastOpenedAt: row.isNull(at: 2) ? nil : Date(timeIntervalSince1970: row.double(at: 2)),
-                favoritedAt: row.isNull(at: 3) ? nil : Date(timeIntervalSince1970: row.double(at: 3)),
-                positionUpdatedAt: row.isNull(at: 4) ? nil : Date(timeIntervalSince1970: row.double(at: 4)),
-                updatedAt: Date(timeIntervalSince1970: row.double(at: 5)),
-                isFavorite: row.int(at: 6) != 0,
-                ckRecordId: row.string(at: 7)
-            )
-        }) ?? []
-
-        entriesByBookId = Dictionary(uniqueKeysWithValues: entries.map { ($0.bookId, $0) })
-
-        historyOrder = (try? _db.fetch(query: "SELECT book_id FROM history_order ORDER BY position;") { row -> Int in
-            row.int(at: 0)
-        }) ?? []
+        let data = HistoryDatabaseManager.shared.loadFromDatabase()
+        entriesByBookId = Dictionary(uniqueKeysWithValues: data.entries.map { ($0.bookId, $0) })
+        historyOrder = data.historyOrder
     }
 
     // MARK: - Core Operations
@@ -299,8 +148,8 @@ class HistoryViewModel: ViewModelBase, ObservableObject {
         }
 
         pruneOrphanedEntries()
-        upsertEntry(entry)
-        saveHistoryOrder()
+        HistoryDatabaseManager.shared.upsertEntry(entry)
+        HistoryDatabaseManager.shared.saveHistoryOrder(historyOrder)
         loadBooksData()
 
         CloudKitSyncManager.shared.uploadHistory(entries: [entry])
@@ -317,10 +166,10 @@ class HistoryViewModel: ViewModelBase, ObservableObject {
             entriesByBookId[bookId] = entry
 
             // Hanya simpan ke DB — tidak reload UI library (tidak ada perubahan visible)
-            upsertEntry(entry)
+            HistoryDatabaseManager.shared.upsertEntry(entry)
 
             if let ckId = entry.ckRecordId {
-                addPendingSync(ckRecordId: ckId, operation: "upload")
+                HistoryDatabaseManager.shared.addPendingSync(ckRecordId: ckId, operation: "upload")
             }
             CloudKitSyncManager.shared.uploadHistory(entries: [entry])
         } else {
@@ -354,7 +203,7 @@ class HistoryViewModel: ViewModelBase, ObservableObject {
         }
 
         entriesByBookId[bookId] = entry
-        upsertEntry(entry)
+        HistoryDatabaseManager.shared.upsertEntry(entry)
         loadBooksData()
 
         CloudKitSyncManager.shared.uploadHistory(entries: [entry])
@@ -368,19 +217,19 @@ class HistoryViewModel: ViewModelBase, ObservableObject {
                 entry.lastOpenedAt = nil
                 entry.updatedAt = Date()
                 entriesByBookId[bookId] = entry
-                upsertEntry(entry)
-                saveHistoryOrder()
+                HistoryDatabaseManager.shared.upsertEntry(entry)
+                HistoryDatabaseManager.shared.saveHistoryOrder(historyOrder)
                 loadBooksData()
                 CloudKitSyncManager.shared.uploadHistory(entries: [entry])
             } else {
                 // Entry dihapus total — delete di CloudKit, tidak perlu upload
                 let ckId = entry.ckRecordId
                 entriesByBookId.removeValue(forKey: bookId)
-                deleteEntry(bookId: bookId)
-                saveHistoryOrder()
+                HistoryDatabaseManager.shared.deleteEntry(bookId: bookId)
+                HistoryDatabaseManager.shared.saveHistoryOrder(historyOrder)
                 loadBooksData()
                 if let ckId {
-                    addPendingSync(ckRecordId: ckId, operation: "delete")
+                    HistoryDatabaseManager.shared.addPendingSync(ckRecordId: ckId, operation: "delete")
                     pendingCloudKitDeletes.insert(ckId)
 
                     deleteDebounceTask?.cancel()
@@ -411,18 +260,18 @@ class HistoryViewModel: ViewModelBase, ObservableObject {
                     entry.lastOpenedAt = nil
                     entry.updatedAt = Date()
                     entriesByBookId[bookId] = entry
-                    upsertEntry(entry)
+                    HistoryDatabaseManager.shared.upsertEntry(entry)
                 } else {
                     if let ckId = entry.ckRecordId {
                         ckIdsToDelete.append(ckId)
                     }
                     entriesByBookId.removeValue(forKey: bookId)
-                    deleteEntry(bookId: bookId)
+                    HistoryDatabaseManager.shared.deleteEntry(bookId: bookId)
                 }
             }
         }
 
-        saveHistoryOrder()
+        HistoryDatabaseManager.shared.saveHistoryOrder(historyOrder)
         loadBooksData()
 
         if !ckIdsToDelete.isEmpty {
@@ -446,7 +295,7 @@ class HistoryViewModel: ViewModelBase, ObservableObject {
         }
         for bookId in toRemove {
             entriesByBookId.removeValue(forKey: bookId)
-            deleteEntry(bookId: bookId)
+            HistoryDatabaseManager.shared.deleteEntry(bookId: bookId)
         }
     }
 
@@ -482,7 +331,7 @@ class HistoryViewModel: ViewModelBase, ObservableObject {
 
             for bookId in bookIdsToDelete {
                 entriesByBookId.removeValue(forKey: bookId)
-                deleteEntry(bookId: bookId)
+                HistoryDatabaseManager.shared.deleteEntry(bookId: bookId)
                 historyOrder.removeAll(where: { $0 == bookId })
                 didChange = true
             }
@@ -495,12 +344,12 @@ class HistoryViewModel: ViewModelBase, ObservableObject {
 
                     if remoteModified > localModified {
                         entriesByBookId[remoteEntry.bookId] = remoteEntry
-                        upsertEntry(remoteEntry)
+                        HistoryDatabaseManager.shared.upsertEntry(remoteEntry)
                         didChange = true
                     }
                 } else {
                     entriesByBookId[remoteEntry.bookId] = remoteEntry
-                    upsertEntry(remoteEntry)
+                    HistoryDatabaseManager.shared.upsertEntry(remoteEntry)
                     didChange = true
                 }
             }
@@ -514,7 +363,7 @@ class HistoryViewModel: ViewModelBase, ObservableObject {
                 historyOrder = Array(sortedIds.prefix(maxHistoryCount))
 
                 pruneOrphanedEntries()
-                saveHistoryOrder()
+                HistoryDatabaseManager.shared.saveHistoryOrder(historyOrder)
                 loadBooksData()
             }
         }
@@ -529,115 +378,6 @@ class HistoryViewModel: ViewModelBase, ObservableObject {
         return true
     }
 
-    // MARK: - Pending Sync (SQLite)
-
-    func addPendingSync(ckRecordId: String, operation: String) {
-        guard let _db else { return }
-        do {
-            if operation == "upload" {
-                let count = try _db.fetch(
-                    query: "SELECT COUNT(*) FROM sync_pending WHERE ck_record_id = ? AND operation = 'delete';",
-                    parameters: [ckRecordId]
-                ) { $0.int64(at: 0) }.first ?? 0
-                if count > 0 { return } // Delete wins
-            } else if operation == "delete" {
-                try _db.execute(
-                    query: "DELETE FROM sync_pending WHERE ck_record_id = ? AND operation = 'upload';",
-                    parameters: [ckRecordId]
-                )
-            }
-            try _db.execute(
-                query: "INSERT OR REPLACE INTO sync_pending (ck_record_id, operation, queued_at) VALUES (?, ?, ?);",
-                parameters: [ckRecordId, operation, Int64(Date().timeIntervalSince1970)]
-            )
-        } catch {
-            #if DEBUG
-            print("HistoryViewModel: addPendingSync error: \(error)")
-            #endif
-        }
-    }
-
-    func removePendingSync(ckRecordIds: [String]) {
-        guard let _db, !ckRecordIds.isEmpty else { return }
-        let placeholders = ckRecordIds.map { _ in "?" }.joined(separator: ",")
-        try? _db.execute(
-            query: "DELETE FROM sync_pending WHERE ck_record_id IN (\(placeholders));",
-            parameters: ckRecordIds
-        )
-    }
-
-    func fetchPendingSync(operation: String) -> [String] {
-        guard let _db else { return [] }
-        return (try? _db.fetch(
-            query: "SELECT ck_record_id FROM sync_pending WHERE operation = ? ORDER BY queued_at ASC;",
-            parameters: [operation]
-        ) { $0.string(at: 0) ?? "" }) ?? []
-    }
-
-    // MARK: - Migration: UserDefaults → SQLite
-
-    private func migrateFromUserDefaultsIfNeeded() {
-        guard !UserDefaults.standard.bool(forKey: migrationFlag) else { return }
-        guard let _db else { return }
-
-        // Migrate main entries
-        if let data = UserDefaults.standard.data(forKey: legacyStorageKey),
-           let stored = try? JSONDecoder().decode(StoredReadingEntries.self, from: data)
-        {
-            do {
-                try _db.transaction {
-                    for entry in stored.entries {
-                        upsertEntry(entry)
-                    }
-                    // Preserve exact history order
-                    for (position, bookId) in stored.historyOrder.enumerated() {
-                        try exec("INSERT OR REPLACE INTO history_order (position, book_id) VALUES (?, ?);", parameters: [position, bookId])
-                    }
-                }
-            } catch {
-                #if DEBUG
-                print("HistoryViewModel: Migration failed: \(error)")
-                #endif
-                return // Don't mark as migrated if failed
-            }
-        }
-
-        // Migrate pending sync
-        if let upData = UserDefaults.standard.data(forKey: legacyPendingUploadsKey),
-           let upList = try? JSONDecoder().decode([String].self, from: upData)
-        {
-            let now = Int64(Date().timeIntervalSince1970)
-            for ckId in upList {
-                try? _db.execute(
-                    query: "INSERT OR REPLACE INTO sync_pending (ck_record_id, operation, queued_at) VALUES (?, 'upload', ?);",
-                    parameters: [ckId, now]
-                )
-            }
-        }
-
-        if let delData = UserDefaults.standard.data(forKey: legacyPendingDeletesKey),
-           let delList = try? JSONDecoder().decode([String].self, from: delData)
-        {
-            let now = Int64(Date().timeIntervalSince1970)
-            for ckId in delList {
-                try? _db.execute(
-                    query: "INSERT OR REPLACE INTO sync_pending (ck_record_id, operation, queued_at) VALUES (?, 'delete', ?);",
-                    parameters: [ckId, now]
-                )
-            }
-        }
-
-        // Mark migrated and delete old UserDefaults data
-        UserDefaults.standard.set(true, forKey: migrationFlag)
-        UserDefaults.standard.removeObject(forKey: legacyStorageKey)
-        UserDefaults.standard.removeObject(forKey: legacyPendingUploadsKey)
-        UserDefaults.standard.removeObject(forKey: legacyPendingDeletesKey)
-
-        #if DEBUG
-        print("HistoryViewModel: Successfully migrated from UserDefaults to SQLite")
-        #endif
-    }
-
     // MARK: - KVS Migration (Legacy)
 
     func backfillCloudKitFieldsIfNeeded(completion: (([ReadingEntry]) -> Void)? = nil) {
@@ -650,7 +390,7 @@ class HistoryViewModel: ViewModelBase, ObservableObject {
                 updated.ckRecordId = String(bookId)
                 entriesByBookId[bookId] = updated
                 backfilled.append(updated)
-                upsertEntry(updated)
+                HistoryDatabaseManager.shared.upsertEntry(updated)
                 didChange = true
             }
         }
@@ -687,7 +427,7 @@ class HistoryViewModel: ViewModelBase, ObservableObject {
                     var migrated = entry
                     migrated.ckRecordId = String(entry.bookId)
                     entriesByBookId[entry.bookId] = migrated
-                    upsertEntry(migrated)
+                    HistoryDatabaseManager.shared.upsertEntry(migrated)
                 }
             }
 
@@ -697,7 +437,7 @@ class HistoryViewModel: ViewModelBase, ObservableObject {
                 }
             }
 
-            saveHistoryOrder()
+            HistoryDatabaseManager.shared.saveHistoryOrder(historyOrder)
 
             // Upload semua entry hasil migrasi ke CloudKit (satu kali, batch)
             let migratedEntries = Array(entriesByBookId.values).filter { $0.ckRecordId != nil }
@@ -729,9 +469,9 @@ class HistoryViewModel: ViewModelBase, ObservableObject {
         }
 
         // Update DB: delete old, insert new
-        deleteEntry(bookId: oldId)
-        upsertEntry(migrated)
-        saveHistoryOrder()
+        HistoryDatabaseManager.shared.deleteEntry(bookId: oldId)
+        HistoryDatabaseManager.shared.upsertEntry(migrated)
+        HistoryDatabaseManager.shared.saveHistoryOrder(historyOrder)
 
         // Hapus entry lama dari CloudKit
         if let oldCkId = entry.ckRecordId {
